+++ app/src/main/java/org/sopt/sample/config/BaseActivity.kt
@@ -21,8 +21,8 @@ abstract class BaseActivity<B : ViewBinding>(private val inflate: (LayoutInflate
         Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
     }
 
-    fun showCustomSnackBar(message: String) {
-        Snackbar.make(binding.root, message, Snackbar.LENGTH_SHORT).show()
+    fun CustomSnackBar(message: String): Snackbar {
+        return Snackbar.make(binding.root, message, Snackbar.LENGTH_SHORT)
     }
