+++ app/src/main/res/layout/activity_signin.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".src.SignInActivity">
+    tools:context=".src.login.SignInActivity">
 
     <TextView
         android:id="@+id/signIn_title_tv"
